requirements.txt
asgiref==3.2.10

requests==2.26.0

django==2.2.16
django-filter==21.1

djangorestframework==3.12.4

django-cors-headers==3.11.0

django-extra-fields==3.0.2

djangorestframework-simplejwt==4.7.2 

djoser==2.1.0

gunicorn==20.0.4
psycopg2-binary==2.8.6

Pillow==9.3.0

PyJWT==2.1.0
python-dotenv==0.20.0
pytz==2020.1
sqlparse==0.3.1 

reportlab==3.6.11